Fixed error in location of test tools
